backend/requirements.txt
django>=6.0
djangorestframework>=3.17
django-cors-headers>=4.9
python-dotenv>=1.2
groq>=0.9
